+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- [`init`] Now `cargo compete init` creates a `.gitignore` when it is missing.
 - [`new`, `open`] CWD for `jq` is set to the workspace root.
 
 ## [0.4.0] - 2020-08-09Z